refactor(docs): kbd dx
