+++ apps/docs/content/components/kbd/keys.ts
@@ -1,14 +1,4 @@
-const App = `import {Kbd} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex gap-4">
-      <Kbd keys={["command"]}>K</Kbd>
-      <Kbd keys={["command", "shift"]}>N</Kbd>
-      <Kbd keys={["option", "command"]}>P</Kbd>
-    </div>  
-  );
-}`;
+import App from "./keys.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
