+++ apps/docs/content/components/kbd/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {Kbd} from "@nextui-org/react";
+
+export default function App() {
+  return <Kbd keys={["command"]}>K</Kbd>;
+}
