+++ apps/docs/content/components/kbd/usage.ts
@@ -1,10 +1,4 @@
-const App = `import {Kbd} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Kbd keys={["command"]}>K</Kbd>
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
